CI refactor (#62)
Why:
We should try to minimize the size of CI job files

What:
I have created a composite action file for each OS we are testing on, so
it can be reused in unit_test.yml and foss.yml.
I have modified the unit_test.yml file to use these composite actions.

Addresses:
#55 

Depends on:
#51