Refactor config rule (#104)
Why:
We want to handle configurations for both the monolithic and `per_file`
rule.
I want to use the same implementation for both, and don't want to either
repeat code or have cyclic includes.

What:
I have created a file `codechecker_config.bzl` in which everything
related to the rule codechecker_config lives.
I have included this rule in `codechecker.bzl` and used it.

Addresses:
-
(In the long run: #30 )